+++ R/schema.R
@@ -459,6 +459,14 @@ set_not_null_constraints_on_fk_cols <- function(schema,
 #' @param compress A boolean that tells whether to compress `db_table_name`
 #'                 in the database after uploading.
 #'                 Default is `FALSE`.
+#' @param round_double_columns A boolean that tells whether to
+#'                             round double-precision columns in `.df`.
+#'                             Default is `FALSE`.
+#' @param digits An integer that tells the number of significant digits.
+#'               `digits` has an effect only when `round_double_columns` is `TRUE`.
+#'               Default is `15`, which should
+#'               eliminate any numerical precision errors
+#'               for [compress_rows()].
 #' @param index_map A list of 2 or more data frames that represent the
 #'                  mappings from inboard row and column indices in the database
 #'                  to outboard row and column names in the memory
@@ -498,6 +506,8 @@ pl_upsert <- function(.df,
                       in_place = FALSE,
                       encode_fks = TRUE,
                       compress = FALSE,
+                      round_double_columns = FALSE,
+                      digits = 15,
                       index_map = list(fk_parent_tables[[IEATools::row_col_types$industry]],
                                        fk_parent_tables[[IEATools::row_col_types$product]],
                                        fk_parent_tables[[IEATools::row_col_types$other]]) |>
@@ -557,6 +567,12 @@ pl_upsert <- function(.df,
       encode_fks(db_table_name = db_table_name,
                  schema = schema,
                  fk_parent_tables = fk_parent_tables)
+  }
+
+  # Round double-precision columns, if desired
+  if (round_double_columns) {
+    df_to_upsert <- df_to_upsert |>
+      round_double_cols(digits = digits)
   }
 
   # Perform the upload.